start iteration 3
Modified the application startup mode (application factory mode). A blueprint is added to the application to make the entire application modular and easy to expand functions. Added the ability for users to choose avatars to the application (implemented through API). A page for setting personal information has been added to the application (user name and email can be modified).